+++ docs/CHANGELOG.md
@@ -13,6 +13,10 @@ The versioning system is basically the following:
   - If there's a really "big" change.
 - As mentioned, the "major" version is 0.
 
+## main
+
+- Allow filtering out (i.e. ignoring) all diagnostics on certain files with the `milletDiagnosticsFilter all` ML Basis annotation.
+
 ## v0.7.6
 
 - Show unsolved equality type variables with two `?`.